Removed Audio Demos and integrated into play screed
