+++ App.tsx
@@ -57,8 +57,8 @@ export default function App() {
 
   return !resourcesLoaded ? null : (
     <View style={styles.container} onLayout={onLayoutRootView}>
-      {/* <RootNavigation /> */}
-      <AudioDemos />
+      {/* <AudioDemos /> */}
+      <RootNavigation />
     </View>
   );
 }
