+++ package.json
@@ -44,8 +44,8 @@
     "react-native-safe-area-context": "4.5.0",
     "react-native-screens": "~3.20.0",
     "react-native-svg": "13.4.0",
-    "react-navigation": "^4.4.4",
     "react-native-web": "~0.18.9",
+    "react-navigation": "^4.4.4",
     "uri": "^0.1.0"
   },
   "devDependencies": {
